azure/requirements.txt
aiohttp
botbuilder-integration-aiohttp>=4.14.0
langchain
google-cloud-aiplatform
google-cloud-discoveryengine